requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyMySQL==0.9.3
PyYAML==5.3.1
Flask==1.1.1
boto3==1.13.3
numpy==1.18.1
pandas==1.0.3
requests==2.21.0
scikit-learn==0.22.1
xgboost==1.0.2